refactor: organize files into purpose-based folder
