+++ app/src/main/java/com/example/movieapp/di/UseCaseModule.kt
@@ -2,12 +2,12 @@ package com.example.movieapp.di
 
 import com.example.domain.repository.AuthRepository
 import com.example.domain.repository.MovieRepository
-import com.example.domain.usecase.GetMovieUseCase
-import com.example.domain.usecase.GetPopularMoviesUseCase
-import com.example.domain.usecase.LoginUseCase
-import com.example.domain.usecase.impl.GetMovieUseCaseImpl
-import com.example.domain.usecase.impl.GetPopularMoviesUseCaseImpl
-import com.example.domain.usecase.impl.LoginUseCaseImpl
+import com.example.domain.usecase.getmovie.GetMovieUseCase
+import com.example.domain.usecase.getmovie.GetMovieUseCaseImpl
+import com.example.domain.usecase.getpopularmovie.GetPopularMoviesUseCase
+import com.example.domain.usecase.getpopularmovie.GetPopularMoviesUseCaseImpl
+import com.example.domain.usecase.login.LoginUseCase
+import com.example.domain.usecase.login.LoginUseCaseImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
