+++ domain/src/main/java/com/example/domain/usecase/getpopularmovie/GetPopularMoviesUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.domain.usecase
+package com.example.domain.usecase.getpopularmovie
 
 import androidx.paging.PagingData
 import com.example.domain.entity.MovieEntity
