+++ moviecomposeapp/src/main/java/com/example/moviecomposeapp/screen/login/LoginViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.data.network.DispatcherProvider
 import com.example.domain.either.Either
-import com.example.domain.usecase.LoginUseCase
+import com.example.domain.usecase.login.LoginUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
